ml/passport/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
paddleocr==2.7.3
pillow==10.4.0
numpy==1.26.4
